Ensuring tests pick up declaration files.
