+++ package.json
@@ -20,7 +20,7 @@
     "tsc": "tsc --noEmit",
     "build:ts": "tsc",
     "test": "yarn test:unit && yarn test:integration && yarn test:e2e",
-    "test:unit": "mocha --config .mocharc.json",
+    "test:unit": "TS_NODE_PROJECT=./tsconfig.test.json TS_NODE_FILES=true mocha --config .mocharc.json",
     "test:integration": "yarn redis:start && mocha --config .mocharc.integration.json && yarn redis:stop",
     "test:e2e": "./scripts/run-playwright.sh",
     "coverage": "TS_NODE_TRANSPILE_ONLY=true NODE_OPTIONS='--import ts-node/esm' c8 mocha -r tsconfig-paths/register 'tests/unit/**/*.spec.ts'",
